+++ finagle-mux/src/test/java/com/twitter/finagle/mux/RequestsCompilationTest.java
@@ -1,0 +1,17 @@
+package com.twitter.finagle.mux;
+
+import org.junit.Test;
+
+import com.twitter.io.Buf;
+import com.twitter.io.Bufs;
+
+public class RequestsCompilationTest {
+
+  private Buf buf = Bufs.EMPTY;
+
+  @Test
+  public void makeWithPayload() {
+    Request request = Requests.make(buf);
+  }
+
+}